+++ federation-integration-testsuite-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG for `federation-integration-testsuite-js`
 
+## 2.5.1
+
+## 2.5.0
+
+## 2.4.10
+
 ## 2.4.9
 
 ## 2.4.8